Interface.TeX

Description

TeX interface

Documentation

class Typeset a where

Typeset is a class of types whose terms can be translated into a TeX source (in Data.Text.Lazy). toTeX method translates a Typeset term into a TeX source (in Data.Text.Lazy).

Minimal complete definition

toTeX

Methods

toTeX :: a -> Text

scaleboxsize :: Text -> Text

determines a scale of a parsing result (a syntactic derivation or an SR) given to x in scalebox{x}{...}, according to the length of the sentence to be parsed.