Parser.CCG

Description

Syntactic categories, syntactic features and combinatory rules of Combinatory Categorial Grammar.

Types

data Node

A node in CCG derivation tree.

Constructors

Node

Fields

rs :: RuleSymbol

The name of the rule

pf :: Text

The phonetic form

cat :: Cat

The syntactic category (in CCG)

sem :: Preterm

The semantic representation (in DTS)

sig :: Signature

Signature

daughters :: [Node]

The daughter nodes

score :: Rational

The score (between 0.00 to 1.00, larger the better)

source :: Text

The source of the lexical entry

data RuleSymbol

The name of the CCG rule to derive the node.

Constructors

LEX
A lexical item

EC
An empty category

FFA
Forward function application rule.

BFA
Backward function application rule

FFC1
Forward function composition rule 1

BFC1
Backward function composition rule 1

FFC2
Forward function composition rule 2

BFC2
Backward function composition rule 2

FFC3
Forward function composition rule 3

BFC3
Backward function composition rule 3

FFCx1
Forward function crossed composition rule 1

FFCx2
Forward function crossed composition rule 2

FFSx
Forward function crossed substitution rule

COORD
Coordination rule

PAREN
Parenthesis rule

WRAP
Wrap rule

DC
Dynamic conjunction rule

DREL
Discourse Relation rule

data Cat

Syntactic categories of

Constructors

S [Feature]
S

NP [Feature]
NP

N
N

Sbar [Feature]
S bar

CONJ
CON

LPAREN
A category for left parentheses

RPAREN
A category for right parentheses

SL Cat Cat
X/Y

BS Cat Cat
X\Y

T Bool Int Cat
Category variables, where Int is an index, Cat is a restriction for its head.

data Feature

Syntactic features of

Constructors

F [FeatureValue]
Syntactic feature

SF Int [FeatureValue]
Shared syntactic feature (with an index)

data FeatureValue

Values of syntactic features of Japanese

Constructors

V5k

V5s

V5t

V5n

V5m

V5r

V5w

V5g

V5z

V5b

V5IKU

V5YUK

V5ARU

V5NAS

V5TOW

V1

VK

VS

VSN

VZ

VURU

Aauo

Ai

ANAS

ATII

ABES

Nda

Nna

Nno

Ntar

Nni

Nemp

Nto

Exp

Stem

UStem

NStem

Neg

Cont

Term

Attr

Hyp

Imper

Pre

NTerm

NegL

TeForm

NiForm

EuphT

EuphD

ModU

ModD

ModS

ModM

VoR

VoS

VoE

P

M

Nc

Ga

O

Ni

To

Niyotte

No

ToCL

YooniCL

Decl

Text and TeX interface

showScore :: Node -> Text

shows the score of the given node

Tests

isBaseCategory :: Cat -> Bool

A test to check if a given category is a base category (i.e. not a functional category nor a category variable).

isBunsetsu :: Cat -> Bool

A test to check if a given category is the one that can appear on the left adjacent of a punctuation.

Combinatory Rules

unaryRules :: Node -> [Node] -> [Node]

The function to apply all the unaryRules to a CCG node.

binaryRules :: Node -> Node -> [Node] -> [Node]

The function to apply all the binary rules to a given pair of CCG nodes.

coordinationRule :: Node -> Node -> Node -> [Node] -> [Node]

Coordination rule.

parenthesisRule :: Node -> Node -> Node -> [Node] -> [Node]

Parenthesis rule.

for debugging

unifyCategory

Arguments

:: Assignment Cat
A category assignment function

-> Assignment [FeatureValue]
A feature assignment function

-> [Int]
A list of banned indices (unification to which is banned, preventing cyclic unification)

-> Cat
A first argument

-> Cat
A second argument

-> Maybe (Cat, Assignment Cat, Assignment [FeatureValue])

unifies two syntactic categories (Cat) and returns a unified syntactic category, under a given category assignment and a given feature assignment.

unifyWithHead

Arguments

:: Assignment Cat

-> Assignment [FeatureValue]
A feature assignment function.

-> [Int]
A list of indices

-> Cat
A first argument

-> Cat
A second argument

-> Maybe (Cat, Assignment Cat, Assignment [FeatureValue])

unifies a cyntactic category c1 (in `T True i c1`) with the head of c2, under a given feature assignment.

Partial Parsing

wrapNode :: Node -> Node

receives a node and returns an Sbar node, whose SR is obtained by existentially quantifying all the missing arguments of the SR of a given node.

conjoinNodes :: Node -> Node -> Node

receives a pair of Sbar nodes, and returns a node whose SR is their conjunction. Note that this function does not check if given nodes are of Sbar.