Parser.Japanese.Lexicon

Description

A Japanese CCG lexicon.

Documentation

type LexicalItems = [Node]

Lexicon consists of a set of CCG Nodes

lookupLexicon :: Text -> LexicalItems -> [Node]

This function takes a word and a lexicon and returns a set of CCG lexical entries whose PF is that word.

setupLexicon :: Text -> IO LexicalItems

This function takes a sentence and returns a numeration needed to parse that sentence, i.e., a union of

emptyCategories :: [Node]

A list of empty categories (i.e. lexical items whose PF is empty).

myLexicon :: [Node]

A list of (mostly functional) lexical items extracted from Bekki (2010).