Parser.Japanese.Templates

Description

A set of lexical templates for building a lexicon.

Templates for lexical items

lexicalitem

Arguments

:: Text
A phonetic form

-> Text
A source of a lexical item, such as a number in the CCG textbook.

-> Integer
A score (0 to 100)

-> Cat
A syntactic category

-> (Preterm, Signature)
A semantic representation (in DTS) and a list of signatures

-> Node

defines a lexical item.

Templates for CCG syntactic features

defS :: [FeatureValue] -> [FeatureValue] -> Cat

Category S with the default feature setting (mainly for stems).

verb :: [FeatureValue]

A set of conjugation forms of JP verbs.

adjective :: [FeatureValue]

A set of conjugation forms of JP adjectives.

nomPred :: [FeatureValue]

A set of conjugation forms of JP nominal predicates.

nonStem :: [FeatureValue]

modifiableS :: Cat

anyPos :: [FeatureValue]

All conjugation forms, i.e. verb ++ adjective ++ nomPred

m5 :: [Feature]

pmmmm :: [Feature]

mpmmm :: [Feature]

mmpmm :: [Feature]

mmmpm :: [Feature]

mppmm :: [Feature]

Templates for DTS representations

id :: Preterm

Lam x.x

entity :: Preterm

event :: Preterm

state :: Preterm

verbCat

Arguments

:: Text
a case frame (e.g. "ガヲニ")

-> [FeatureValue]
a part-of-speech feature value

-> [FeatureValue]
a conjugational feature value

-> Cat

verbSR :: Text -> Preterm -> Text -> (Preterm, Signature)

verbSR i op i==1 -> SNP: x.c.(e:event)Xop(e,x)X(ce) i==2 -> SNPNP: y.x.c.(e:event)X(op(e,x,y)X(ce) i==3 -> SNPNPNP: z.y.x.c.(e:event)X(op(e,x,y,z)X(ce) ...

predSR :: Int -> Text -> (Preterm, Signature)

SNP: x.c.(s:state)Xop(s,x)X(ce)

nPlaceVerbType

Arguments

:: Preterm
Eventuarlity

-> Text
A case frame

-> Preterm

nPlaceEventType :: Int -> Preterm

nPlaceStateType :: Int -> Preterm

nPlacePredType :: Int -> Preterm

properNameSR :: Text -> (Preterm, Signature)

NP:

commonNounSR :: Text -> (Preterm, Signature)

N:

generalizedQuantifierSR :: Text -> (Preterm, Signature)

GQ:

intensionalEvent :: Int -> Text -> (Preterm, Signature)

NN, N/N modifierSR :: T.Text -> (Preterm,Signature) modifierSR op = ((Lam (Lam (Lam (App (App (Var 2) (Var 1)) (Lam (Sigma (App (Con op) (Var 0)) (App (Var 2) (Var 1)))))))), [(op, nPlacePredType 1)])

>>> S\NP\(S\NP):    \p.\x.\c.op(x,\z.(pz)c)
>>> S\NP\NP\(S\NP): \p.\y.\x.\c.op(x,\z.((py)z)c)

intensionalState :: Int -> Text -> (Preterm, Signature)

modalSR :: Text -> (Preterm, Signature)

SS, S/S: p.c.op (pc)

mannerAdverb :: Text -> (Preterm, Signature)

T/T: p.v.c.pv(e.(op e) X ce)

eventModifier :: Text -> (Preterm, Signature)

SS: p.c.p(e.(op e) X ce)

nominalModifier :: Text -> (Preterm, Signature)

N/N, NN: n.x.c. (nx (s.(op x) × cs))

negOperator :: (Preterm, Signature)

negation operator SS: p.c.not (pc)

negOperator2 :: (Preterm, Signature)

negation operator 2 SNP(SNP): p.x.c.not ((px)c)

argumentCM :: (Preterm, Signature)

argument case marker T/(TNP[cm])NP[nc]: x.p.px

adjunctCM :: Text -> (Preterm, Signature)

adjunct case marker S1/S1NP[nc]: x.p.c.p (e.op(e,x) X ce)

adjunctNM :: Text -> (Preterm, Signature)

adjunct nominal modifier NN(TTNP[nc]): p.n.x.c.(nx (s.p(z.op(s,z)) X cs)

andSR :: (Preterm, Signature)

orSR :: (Preterm, Signature)

conjunctionSR :: Text -> (Preterm, Signature)