DTS.DTT

Description

Dependent Type Theory (in de Bruijn index)

Types

data Preterm

Preterms of Underspecified Dependent Type Theory (DTT).

Constructors

Var Int
Variables

Con Text
Constant symbols

Type
The sort "type"

Kind
The sort "kind"

Pi Preterm Preterm
Dependent function types (or Pi types)

Not Preterm
Negations

Lam Preterm
Lambda abstractions

App Preterm Preterm
Function Applications

Sigma Preterm Preterm
Dependent product types (or Sigma types)

Pair Preterm Preterm
Pairs

Proj Selector Preterm
(First and second) Projections

Unit
The unit term (of type Top)

Top
The top type

Bot
The bottom type

Nat
Natural number type (Nat)

Zero
0 (of type Nat)

Succ Preterm
The successor function

Natrec Preterm Preterm Preterm
natrec

Eq Preterm Preterm Preterm
Intensional equality types

Refl Preterm Preterm
refl

Idpeel Preterm Preterm
idpeel

DRel Int Text Preterm Preterm
Discourse relations

data Selector

Proj Fst m is the first projection of m, while Proj Snd m is the second projection of m.

Constructors

Fst

Snd

type Signature = [(Text, Preterm)]

A type of an element of a type signature, that is, a list of pairs of a preterm and a type. ex. [entity:type, state:type, event:type, student:entity->type]

toUDTT :: Preterm -> Preterm

DTT to UDTT

toDTT :: Preterm -> Preterm

UDTT to DTT