DTS.UDTT

Description

Implementation of Underspecified Dependent Type Theory (Bekki forthcoming).

Types

data Preterm

Preterms of Underspecified Dependent Type Theory (UDTT).

Constructors

Var Int
Variables

Con Text
Constant symbols

Type
The sort "type"

Kind
The sort "kind"

Pi Preterm Preterm
Dependent function types (or Pi types)

Not Preterm
Negations

Lam Preterm
Lambda abstractions

App Preterm Preterm
Function Applications

Sigma Preterm Preterm
Dependent product types (or Sigma types)

Pair Preterm Preterm
Pairs

Proj Selector Preterm
(First and second) Projections

Asp Int Preterm
Underspesified terms

Lamvec Preterm
Lambda abstractions of a variable vector

Appvec Int Preterm
Function applications of a variable vector

Unit
The unit term (of type Top)

Top
The top type

Bot
The bottom type

Nat
Natural number type (Nat)

Zero
0 (of type Nat)

Succ Preterm
The successor function

Natrec Preterm Preterm Preterm
natrec

Eq Preterm Preterm Preterm
Intensional equality types

Refl Preterm Preterm
refl

Idpeel Preterm Preterm
idpeel

DRel Int Text Preterm Preterm
Discourse relations

data Selector

Proj Fst m is the first projection of m, while Proj Snd m is the second projection of m.

Constructors

Fst

Snd

type Signature = [(Text, Preterm)]

A type of an element of a type signature, that is, a list of pairs of a preterm and a type. ex. [entity:type, state:type, event:type, student:entity->type]

toTextDeBruijn :: Preterm -> Text

prints a preterm in text, in the De Bruijn style.

Syntactic Operations

subst :: Preterm -> Preterm -> Int -> Preterm

Substitution of the variable i in a preterm M with a preterm L "subst M L i" = M[L/i]

shiftIndices :: Preterm -> Int -> Int -> Preterm

shiftIndices m d i add d to all the indices that is greater than or equal to i within m (=d-place shift)

addLambda :: Int -> Preterm -> Preterm

addLambda i preterm: the first subroutine for transvec function, which takes an index and a preterm, transforms the latter in a way that the Var/Appvec with an index j that is equal or greater than i Ex. addLambda 1 (Appvec 0 m) = Appvec 1 (addLambda 1 m) addLambda 0 (Appvec 0 m) = Appvec 0 (App () (Var 1))

deleteLambda :: Int -> Preterm -> Preterm

deleteLambda i preterm: the second subroutine for transvec function, which takes an index i and a preterm p, transforms the latter in a way that the i-th variable vector within p is deleted.

replaceLambda :: Int -> Preterm -> Preterm

replaceLambda i preterm: the third subroutine for transvec function, which takes an index i and a preterm p, transforms the latter in a way that the i-th variable vector within p is replaced by a single variable.

Computations

betaReduce :: Preterm -> Preterm

Beta reduction

strongBetaReduce :: Int -> Preterm -> Preterm

strong Beta reduction

sigmaElimination :: Preterm -> Preterm

eliminates nested Sigma constructions from a given preterm

add :: Preterm -> Preterm -> Preterm

adds two preterms (of type Nat).

multiply :: Preterm -> Preterm -> Preterm

multiplies two preterms (of type Nat).

De Bruijn notation from/to Variable-name notation

newtype Indexed a

Indexed monad controls indices to be attached to preterms. Arguments correspond to: u for variables for propositions | x for variables for entities | e for variables for eventualities | indices for asp-operators | indices for DReL operators

Constructors

Indexed

Fields

indexing :: Int -> Int -> Int -> Int -> Int -> Int -> (a, Int, Int, Int, Int, Int, Int)

initializeIndex :: Indexed a -> a

re-assigns sequential indices to all asperands that appear in a given preterm.

fromDeBruijn

Arguments

:: [VarName]
A context (= a list of variable names)

-> Preterm
A preterm in de Bruijn notation

-> Indexed Preterm
A preterm with variable names

translates a preterm in de Bruijn notation into a preterm with variable name.

toDeBruijn

Arguments

:: [VarName]
A context (= a list of variable names)

-> Preterm
A preterm with variable names

-> Preterm
A preterm in de Bruijn notation

translates a preterm with variable name into a preterm in de Bruijn notation.

Judgment

type Context = [Preterm]

A context is a list of preterms

toVerticalMathML :: Context -> Text

prints a context in vertical manner.

printVerticalMathML :: [Preterm] -> IO ()

prints an SR list in vertical manner.

data Judgment

The data type for a judgment

Constructors

Judgment

Fields

context :: Context

A context Gamma in Gamma vdash M:A

term :: Preterm

A term M in Gamma vdash M:A

typ :: Preterm

A type A in Gamma vdash M:A

fromDeBruijnContext :: Context -> Context

translates a context in de Bruijn notation (i.e. [DTS.DependentTypes.Preterm]) into one with variable names (i.e. [(DTS.DTSwithVarName.VarName, DTS.DTSwithVarName.Preterm)]).

fromDeBruijnSRlist :: [Preterm] -> [(VarName, Preterm)]

A list of semantic representations (the first element is for the first sentence)

printProofSearchQuery :: Context -> Preterm -> Text

prints a proof search query in MathML