DTS.UDTTwithName

Description

Underspecified Dependent Type Theory (with variable names).

Documentation

data VarName

A variable name consists of Char (e.g. x) and Int (e.g. 1), which is displayed as $x_{1}$ in TeX and $x1$ in Text.

Constructors

VarName Char Int

data Selector

Proj Fst m is the first projection of m, while Proj Snd m is the second projection of m.

Constructors

Fst

Snd

data Preterm

Preterms of Underspecified Dependent Type Theory (UDTT).

Constructors

Var VarName
Variable

Con Text
Constant symbol

Type
The sort "type"

Kind
The sort "kind"

Pi VarName Preterm Preterm
Dependent function type (or Pi type)

Not Preterm
Negation

Lam VarName Preterm
Lambda abstraction

App Preterm Preterm
Function Application

Sigma VarName Preterm Preterm
Dependent product type (or Sigma type)

Pair Preterm Preterm
Pair

Proj Selector Preterm
(First and second) Projections

Lamvec VarName Preterm
Variable-length lambda abstraction

Appvec VarName Preterm
Variable-length function application

Unit
The unit term (of type Top)

Top
The top type

Bot
The bottom type

Asp Int Preterm
The asperand term (or Underspesified term)

Nat
The natural number type (Nat)

Zero
0 (of type Nat)

Succ Preterm
The successor function

Natrec Preterm Preterm Preterm
natrec

Eq Preterm Preterm Preterm
Intensional equality type

Refl Preterm Preterm
refl

Idpeel Preterm Preterm
idpeel

DRel Int Text Preterm Preterm
Discourse relations

type Context = [(VarName, Preterm)]

A context is a list of pairs of a variable and a preterm.

toVerticalMathML :: Context -> Text

prints a context vertically.

printVerticalMathML :: [(VarName, Preterm)] -> IO ()

prints an SR list vertically.

data Judgment

The data type for a judgment

Constructors

Judgment

Fields

context :: Context

A context Gamma in Gamma vdash M:A

term :: Preterm

A term M in Gamma vdash M:A

typ :: Preterm

A type A in Gamma vdash M:A