Interface.HTML

Description

MathML interface

Documentation

class MathML a where

Math is a class of types whose terms can be translated into a MathML source (in Data.Text.Lazy). toMathML method translates a Typeset term into a TeX source (in Data.Text.Lazy).

Minimal complete definition

toMathML

Methods

toMathML :: a -> Text

htmlHeader4MathML :: String

version='1.0'?...body original source of MathML stylesheet: http://www.w3.org/Math/XSL/mathml.xsl original soruce of MathJax script: http://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML

htmlFooter4MathML :: String

/body/html

startMathML :: Text

xmlns='http://www.w3.org/1998/Math/MathML'

endMathML :: Text

/math