Parser.ChartParser

Description

A left-corner CKY-parser for lexical grammars.

Data structures for CCG derivations

type Chart = Map (Int, Int) [Node]

The type for CYK-charts.

data Node

A node in CCG derivation tree.

Constructors

Node

Fields

rs :: RuleSymbol

The name of the rule

pf :: Text

The phonetic form

cat :: Cat

The syntactic category (in CCG)

sem :: Preterm

The semantic representation (in DTS)

sig :: Signature

Signature

daughters :: [Node]

The daughter nodes

score :: Rational

The score (between 0.00 to 1.00, larger the better)

source :: Text

The source of the lexical entry

Main parsing functions

parse

Arguments

:: Int
The beam width

-> Text
A sentence to be parsed

-> IO Chart
A pair of the resulting CYK-chart and a list of CYK-charts for segments

Main parsing function to parse a Japanees sentence and generates a CYK-chart.

simpleParse :: Int -> Text -> IO [Node]

Simple parsing function to return just the best node for a given sentence

Partial parsing function(s)

data ParseResult

A data type for the parsing result.

Constructors

Full [Node]
when there are at least one node in the topmost box in the chart, returning the nodes.

Partial [Node]
when the parser did not obtain the full result, it collects partially parsed segments from the chart, returning their conjunctions.

Failed
when no box in the chart contains any node.

extractParseResult :: Int -> Chart -> ParseResult

takes a (parse result) chart and returns a list consisting of nodes, partially parsed substrings.